+++ mobile/native/src/lib.rs
@@ -22,3 +22,4 @@ mod orderbook;
     unused_qualifications
 )]
 mod bridge_generated;
+mod destination;